--- src/main/resources/static/cas3v2-api.yml
@@ -1,30 +1,0 @@
-openapi: 3.0.1
-info:
-  version: 1.0.1
-  title: 'Transitional Accommodation Services (CAS3 version 2)'
-servers:
-  - url: /cas3/v2
-paths:
-  /premises/{premisesId}/bookings:
-    get:
-      tags:
-        - Operations on premises
-      summary: Returns all bookings for cas3 premises
-      operationId: premisesPremisesIdBookingsGet
-      parameters:
-        - name: premisesId
-          in: path
-          description: ID of the premises to get bookings for
-          required: true
-          schema:
-            type: string
-            format: uuid
-      responses:
-        200:
-          description: successful operation
-          content:
-            'application/json':
-              schema:
-                type: array
-                items:
-                  $ref: 'cas3-schemas.yml#/components/schemas/Cas3Booking'
